+++ react/components/SkuInventory.tsx
@@ -7,7 +7,11 @@ import { FormattedMessage } from 'react-intl'
 
 const CSS_HANDLES = ['inventoryContainer', 'inventory'] as const
 
-const SkuInventory = () => {
+interface Props {
+  showLabel: boolean
+}
+
+const SkuInventory = ({ showLabel }: Props) => {
   const { sku }: { sku: Item } = useSku()
   const handles = useCssHandles(CSS_HANDLES)
   const inventory: number = sku.sellers
@@ -21,12 +25,32 @@ const SkuInventory = () => {
 
   return (
     <div className={`${handles.inventoryContainer} t-body c-muted-1 lh-copy`}>
-      <span className="bold">
-        <FormattedMessage id="store.sku-list.inventory.title" />:{' '}
-      </span>
+      {showLabel && (
+        <span className="bold">
+          <FormattedMessage id="store.sku-list.inventory.title" />:{' '}
+        </span>
+      )}
       <span className={`${handles.inventory}`}>{inventory}</span>
     </div>
   )
+}
+
+SkuInventory.defaultProps = {
+  showLabel: true,
+}
+
+SkuInventory.schema = {
+  title: 'admin/editor.sku-list.inventory.title',
+  description: 'admin/editor.sku-list.inventory.description',
+  type: 'object',
+  properties: {
+    showLabel: {
+      type: 'boolean',
+      title: 'admin/editor.sku-list.inventory.showLabel.title',
+      default: true,
+      isLayout: false,
+    },
+  },
 }
 
 export default SkuInventory