+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `sku-inventory`
 - Added `showLabel` to `sku-highlights`
 - Indicated unavailability of SKU highlights
+- Added `showLabel` to `sku-inventory`
 
 ### Changed
 - Changed SKU image carousel to new version